+++ v2/managedsettling.go
@@ -65,20 +65,6 @@ func (d *MaxAttemptsRetryDecision) CanRetry(_ error, message *azservicebus.Recei
 	return message.DeliveryCount < d.MaxAttempts
 }
 
-// RetryDelayStrategy can be implemented to provide custom delay retry strategies.
-type RetryDelayStrategy interface {
-	GetDelay(deliveryCount uint32) time.Duration
-}
-
-// ConstantDelayStrategy delays the message retry by the given duration
-type ConstantDelayStrategy struct {
-	Delay time.Duration
-}
-
-func (s *ConstantDelayStrategy) GetDelay(_ uint32) time.Duration {
-	return s.Delay
-}
-
 // ManagedSettlingOptions allows to configure the ManagedSettling middleware
 type ManagedSettlingOptions struct {
 	// Allows to override the built-in error handling logic.
